fixed some tests

+++ jodd-core/src/test/java/jodd/typeconverter/ByteArrayConverterTest.java
@@ -17,23 +17,23 @@ public class ByteArrayConverterTest {
 	public void testArrayConversion() {
 		assertNull(byteArrayConverter.convert(null));
 
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arrb((byte) 1, (byte) 7, (byte) 3)));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arrs((short) 1, (short) 7, (short) 3)));
-		assertEq(arrb((byte) 1, (byte) 0, (byte) 1), byteArrayConverter.convert(arrl(true, false, true)));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arri(1, 7, 3)));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arrl(1, 7, 3)));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arrf(1.99f, 7.99f, 3.22f)));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arrd(1.99, 7.99, 3.22)));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arrs("1", "7", "3")));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(arrs(" 1 ", " 7 ", " 3 ")));
-		assertEq(arrb((byte) 1, (byte) 7, (byte) 3), byteArrayConverter.convert(" 1 ,  7 ,  3 "));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arrb(1, 7, 3)));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arrs(1, 7, 3)));
+		assertEq(arrb(1, 0, 1), byteArrayConverter.convert(arrl(true, false, true)));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arri(1, 7, 3)));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arrl(1, 7, 3)));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arrf(1.99f, 7.99f, 3.22f)));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arrd(1.99, 7.99, 3.22)));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arrs("1", "7", "3")));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(arrs(" 1 ", " 7 ", " 3 ")));
+		assertEq(arrb(1, 7, 3), byteArrayConverter.convert(" 1 ,  7 ,  3 "));
 	}
 
 	@Test
 	public void testNonArrayConversion() {
-		assertEq(arrb((byte) 7), byteArrayConverter.convert(Byte.valueOf((byte) 7)));
-		assertEq(arrb((byte) 7), byteArrayConverter.convert(Integer.valueOf(7)));
-		assertEq(arrb((byte) 7), byteArrayConverter.convert("7"));
+		assertEq(arrb(7), byteArrayConverter.convert(Byte.valueOf((byte) 7)));
+		assertEq(arrb(7), byteArrayConverter.convert(Integer.valueOf(7)));
+		assertEq(arrb(7), byteArrayConverter.convert("7"));
 	}
 
 	private void assertEq(byte[] arr1, byte[] arr2) {
