+++ jodd-core/src/test/java/jodd/typeconverter/ShortArrayConverterTest.java
@@ -17,13 +17,13 @@ public class ShortArrayConverterTest {
 
 		assertNull(shortArrayConverter.convert(null));
 
-		assertEq(arrs((short) 1), shortArrayConverter.convert(Double.valueOf(1)));
-		assertEq(arrs((short) 1, (short) 7, (short) 3), shortArrayConverter.convert(arrs((short) 1, (short) 7, (short) 3)));
-		assertEq(arrs((short) 1, (short) 7, (short) 3), shortArrayConverter.convert(arrb((byte) 1, (byte) 7, (byte) 3)));
-		assertEq(arrs((short) 1, (short) 7, (short) 3), shortArrayConverter.convert(arri(1, 7, 3)));
-		assertEq(arrs((short) 173, (short) 1022), shortArrayConverter.convert(arrs("173", "1022")));
-		assertEq(arrs((short) 173, (short) 1022), shortArrayConverter.convert(arrs(" 173 ", " 1022 ")));
-		assertEq(arrs((short) 173, (short) 10), shortArrayConverter.convert(arro("173", Integer.valueOf(10))));
+		assertEq(arrs(1), shortArrayConverter.convert(Double.valueOf(1)));
+		assertEq(arrs(1, 7, 3), shortArrayConverter.convert(arrs(1, 7, 3)));
+		assertEq(arrs(1, 7, 3), shortArrayConverter.convert(arrb(1, 7, 3)));
+		assertEq(arrs(1, 7, 3), shortArrayConverter.convert(arri(1, 7, 3)));
+		assertEq(arrs(173, 1022), shortArrayConverter.convert(arrs("173", "1022")));
+		assertEq(arrs(173, 1022), shortArrayConverter.convert(arrs(" 173 ", " 1022 ")));
+		assertEq(arrs(173, 10), shortArrayConverter.convert(arro("173", Integer.valueOf(10))));
 	}
 
 	private void assertEq(short[] arr1, short[] arr2) {
